requirements.txt
asgiref==3.5.2
BeautifulSoup4==4.11.1
Django==4.1.2
django-crispy-forms==1.14.0
django-taggit==3.1.0
mysql-connector-python==8.0.31
mysqlclient==2.1.1
Pillow==9.3.0
python-dotenv==0.21.0
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.6
